fix bbox bug & replace default_evalscript with sentinel_utils
