+++ SentinelMosaicTester/pb_tool.cfg
@@ -48,7 +48,7 @@ plugin_path: /Users/natasharavinand/Library/Application Support/QGIS/QGIS3/profi
 
 [files]
 # Python  files that should be deployed with the plugin
-python_files: __init__.py sentinel_mosaic_tester.py sentinel_mosaic_tester_dockwidget.py
+python_files: __init__.py sentinel_mosaic_tester.py sentinel_mosaic_tester_dockwidget.py sentinel_utils.py
 
 # The main dialog file that is loaded (not compiled)
 main_dialog: sentinel_mosaic_tester_dockwidget_base.ui
